# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Person Repository documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/setBankDataPerson": {
            "post": {
                "tags": [
                    "BankData"
                ],
                "summary": "Store the bank account of a person",
                "description": "Storing bank account that belongs to a person",
                "operationId": "StoringBankData",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "person_id",
                                    "bank_name",
                                    "account_number"
                                ],
                                "properties": {
                                    "person_id": {
                                        "type": "number"
                                    },
                                    "bank_name": {
                                        "type": "text"
                                    },
                                    "account_number": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/setDisability": {
            "post": {
                "tags": [
                    "PeopleDisability"
                ],
                "summary": "Store a new disability",
                "description": "Storing disablitity that a person have",
                "operationId": "StoringPeopleDisability",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "person_id",
                                    "disabled_id",
                                    "disabled_level_id"
                                ],
                                "properties": {
                                    "person_id": {
                                        "type": "number"
                                    },
                                    "disabled_id": {
                                        "type": "number"
                                    },
                                    "disabled_level_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/getPersonByDisability/{disability_id}": {
            "get": {
                "tags": [
                    "PeopleDisability"
                ],
                "summary": "Retrieving people by disability",
                "description": "Searching method to get people by their disability",
                "operationId": "retrievingPeopleByDisability",
                "parameters": [
                    {
                        "name": "disability_id",
                        "in": "path",
                        "description": "Id of the disability",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/setPersonByEducation": {
            "post": {
                "tags": [
                    "Education"
                ],
                "summary": "Store the education of a person",
                "description": "Storing education that belongs to a person",
                "operationId": "StoringEducationPerson",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "person_id",
                                    "academic_level_id"
                                ],
                                "properties": {
                                    "person_id": {
                                        "type": "number"
                                    },
                                    "academic_level_id": {
                                        "type": "number"
                                    },
                                    "academic_degree_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/getPersonByEducation": {
            "post": {
                "tags": [
                    "Education"
                ],
                "summary": "Retrieving people by education",
                "description": "Searching method to get people by their education",
                "operationId": "personEducation",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "academic_level_id"
                                ],
                                "properties": {
                                    "academic_level_id": {
                                        "type": "number"
                                    },
                                    "academic_degree_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}